Change type of Item::category from int to ItemCategory

+++ src/elona/item_db.cpp
@@ -136,7 +136,7 @@ void item_db_get_charge_level(const Item& item, int legacy_id)
 
     const auto& info = the_item_db.ensure(legacy_id);
     ichargelevel = info.chargelevel;
-    reftype = info.category;
+    reftype = (int)info.category;
 }
 
 
@@ -153,7 +153,7 @@ void item_db_set_full_stats(Item& item, int legacy_id)
     item.difficulty_of_identification = 0; // Default value
     item.image = info.image;
     fixeditemenc(0) = 0; // Default value
-    reftype = info.category;
+    reftype = (int)info.category;
     reftypeminor = info.subcategory;
 
     switch (legacy_id)
